refactor: use RPCHanlder from source instead of build
